// src/api/generated/models/decisionRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Aegis Backend API
 * OpenAPI spec version: 0.1.0
 */
import type { ReviewDecision } from './reviewDecision.ts';

export interface DecisionRequest {
  decision: ReviewDecision;
  /**
     * @maxLength 1000
     * @nullable
     */
  note?: string | null;
}